refactor(commands): вынести прикладной цикл откликов из _common (#1045)
- commands/supervision.py: ApplyProgress, DurableMutationAttempt, lease+SIGTERM
- commands/apply_service.py: ApplyRunParams, план/провайдеры, волна отклика,
  единый адаптер фиксации результата (_finish_approved_review, _action_status)
- _common.py: только CLI-аргументы/резолв --resume + ре-экспорты
- патч-цели apply-тестов переведены на apply_service; новый unit на параметры

Closes #1045
